feat(lsp): add lspkind plugin so icons are shown in the auto-completion menu
This is a necessary fix since without it you would be only able to use
the tab key for triggering the auto-completion menu or it's selections.
It makes the use of the tab key much limited and the user experience
worse since for example for indentation it would be necessary to use
spaces instead of tabs.